qpackage_deb.bbclass: update pre-process of pkg rename for dpkg
If version is not set in recipe file name and PV is not set in
context, the default PV is set to "git" which is unrecognizable
by Ubuntu platform. Previously, we roughly replace "git" to "0"
during write pkg. This makes version starts with figure and "git"
included(eg.: PV = "0.6.2+git${SRCPV}") in wrong situation.

Set PV to "1.0" if PV is "git". Then the package name of debian
binary and versions in control file are the same.

Change-Id: I1244afa5604fa4b54f95712ae733ad91674c56d0